+++ src/extension.ts
@@ -140,12 +140,17 @@ function getMinOutPath(doc: vscode.TextDocument): string {
     let outNameParts = file.basename.split('.');
 
     outNameParts.pop();
-    if (config.jsPostfix) {
+
+    if (config.jsPostfix && file.languageId === 'javascript') {
+
         outNameParts.push(config.jsPostfix);
-    }
-    if (config.cssPostfix) {
+
+    } else if (config.cssPostfix && file.languageId === 'css') {
+
         outNameParts.push(config.cssPostfix);
+
     }
+
     outNameParts.push(file.extname.replace('.', ''));
     const baseOut = outNameParts.join('.');
 
@@ -218,12 +223,15 @@ function minify(): void {
         extname: path.extname(doc.uri.fsPath),
         dirname: path.dirname(doc.uri.fsPath),
         content: doc.getText(),
-        outpath: getMinOutPath(doc)
+        outpath: getMinOutPath(doc),
+        languageId: doc.languageId
     };
 
     if (file.basename.split('.').length > 2) {
 
-        if (file.basename.split('.')[file.basename.split('.').length - 2] === 'min') {
+        const postfix = file.basename.split('.')[file.basename.split('.').length - 2];
+
+        if ((file.languageId === 'css' && postfix === config.cssPostfix) || (file.languageId === 'javascript' && postfix === config.jsPostfix)) {
             vscode.window.showWarningMessage(`Could not minify ${file.basename}. File already minified.`);
             return;
         }
@@ -380,7 +388,7 @@ function activate(context: vscode.ExtensionContext) {
 
         const da = doc.uri.fsPath.split('.');
 
-        const supported: Array<String> = [
+        const supported: string[] = [
             'javascript',
             'css'
         ];