+++ src/test/workspace/minified/aaa.main.js
@@ -1,1 +1,0 @@
-!function($){class showDivWithText{constructor(param){this.text=param,this.init()}init(){$("<div />").text(this.text).appendTo("body")}}}(jQuery);let a={bs:"fghfghfg",cs:"khfgjhkjfghkljfgjhfghfggfhgf",ds:"gfdgdjghdfjghdfjghdfjlghfd",es:"aaaaaaaaaaaaaaaaaaaaaaaaa"};const b=43,c=(a,b)=>a*b;async function foo(url){return await fetch(url)}console.log(c(a,b));